+++ src/test/java/nextstep/subway/line/LineAcceptanceTest.java
@@ -3,8 +3,6 @@ package nextstep.subway.line;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
-import nextstep.subway.line.dto.LineAcceptanceTestRequest;
-import nextstep.subway.line.dto.LineAcceptanceTestResponse;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.DisplayName;
@@ -13,9 +11,9 @@ import org.junit.jupiter.api.Test;
 import java.util.Arrays;
 import java.util.List;
 
-import static nextstep.subway.line.dto.LineAcceptanceTestRequest.*;
-import static nextstep.subway.line.dto.LineAcceptanceTestResponse.isNoContent;
-import static nextstep.subway.line.dto.LineAcceptanceTestResponse.isStatusOk;
+import static nextstep.subway.line.LineAcceptanceTestRequest.*;
+import static nextstep.subway.line.LineAcceptanceTestResponse.isNoContent;
+import static nextstep.subway.line.LineAcceptanceTestResponse.isStatusOk;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철 노선 관련 기능")
