+++ src/test/java/nextstep/subway/line/LineSectionAcceptanceTestResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.dto;
+package nextstep.subway.line;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
